refactor: 패키지 구조 변경 및 변수 이름 변경에 따른 수정
