+++ src/main/java/chess/dao/GameDao.java
@@ -1,15 +1,13 @@
 package chess.dao;
 
-import chess.service.dto.GameEntity;
-import chess.service.dto.GamesDto;
-import chess.service.dto.StatusDto;
+import java.util.List;
 
 public interface GameDao {
     void update(GameEntity dto);
 
     GameEntity findById(int id);
 
-    GamesDto findAll();
+    List<GameEntity> findAll();
 
     int createGame(String name);
 }
