+++ src/main/java/chess/dao/GameEntity.java
@@ -1,16 +1,13 @@
-package chess.service.dto;
+package chess.dao;
 
 import chess.model.game.ChessGame;
 import chess.model.gamestatus.StatusType;
 
 public class GameEntity {
-    private String status;
-    private String turn;
-    private int id;
+    private final String status;
+    private final String turn;
+    private final int id;
     private String name;
-
-    public GameEntity() {
-    }
 
     public GameEntity(int id, String name, String status, String turn) {
         this.id = id;
@@ -39,21 +36,5 @@ public class GameEntity {
 
     public int getId() {
         return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public void setStatus(String status) {
-        this.status = status;
-    }
-
-    public void setTurn(String turn) {
-        this.turn = turn;
     }
 }
