+++ src/main/java/chess/service/dto/GamesDto.java
@@ -1,5 +1,6 @@
 package chess.service.dto;
 
+import chess.dao.GameEntity;
 import java.util.ArrayList;
 import java.util.List;
