+++ src/test/java/chess/model/board/SquareTest.java
@@ -1,8 +1,10 @@
-package chess.model;
+package chess.model.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import chess.model.File;
+import chess.model.Rank;
 import chess.model.board.Square;
 import chess.model.strategy.move.Direction;
 import chess.model.strategy.move.Distance;
